Rename buffer file to bytes
