+++ src/account.ts
@@ -2,7 +2,7 @@ const assert = require('assert')
 const ethjsUtil = require('ethjs-util')
 const secp256k1 = require('secp256k1')
 import BN = require('bn.js')
-import { toBuffer, addHexPrefix, zeros, bufferToHex, unpad } from './buffer'
+import { toBuffer, addHexPrefix, zeros, bufferToHex, unpad } from './bytes'
 import { keccak, keccak256, rlphash } from './hash'
 
 /**
